I have created the homepage, if anyone would like to add any further details please do. Thank you.
So far I am seperating the CSS files so they can be used amongst multiple pages without interfering with one another.
The main.css should be loaded from every page, and the index.css is only for the index page as it include animations for the first-welcome page. Script.js should also only be loaded for the index.html page as it animates the welcome message.