- index.html Formatting fixes
- team-profile.html
this is just a template page. If there is anything further you would like to be added for templates, please let me know and I'll happy support it. Cheers.

- main.css
Found an element which was still in index.css and not in main.css, transferred it across for suitability. May end up fixing further.